refactor: move schemas to separate file
